+++ rolldown.config.ts
@@ -1,0 +1,7 @@
+import { defineConfig } from "@bemedev/dev-utils/rolldown";
+
+export default defineConfig.bemedev({
+  declarationMap: true,
+  ignoresJS: "**/*.example.ts",
+  excludesTS: ["./*.ts"],
+}) as any;