+++ src/__tests__/actions.machine.test-d.ts
@@ -1,0 +1,7 @@
+import type { PrivateContextFrom } from "@bemedev/app";
+
+expectTypeOf<
+  PrivateContextFrom<typeof import("./actions.machine").default>
+>().toEqualTypeOf<{
+  count: number;
+}>();