+++ src/__tests__/actions.machine.ts
@@ -1,0 +1,14 @@
+import { createMachine, typings } from "@bemedev/app";
+
+export default createMachine(
+  "actions",
+  {
+    initial: "idle",
+    states: { idle: {} },
+  },
+  {
+    pContext: typings.pContext({
+      count: "number",
+    }),
+  },
+);