+++ src/constants/strings.ts
@@ -1,6 +1,0 @@
-export const BIN = 'app-typings';
-
-export const MATCHES = ['**/*.machine.{ts,tsx}'];
-
-export const DESCRIPTION =
-  'Use it to generate type for lib "@bemedev/app-ts"';