+++ src/index.ts
@@ -1,1 +1,6 @@
-export * from './functions';
+#!/usr/bin/env node
+
+import { run } from "cmd-ts";
+import { cli } from "./cli";
+
+run(cli, process.argv.slice(2));