+++ vitest.config.ts
@@ -1,5 +1,5 @@
-import { aliasTs } from '@bemedev/vitest-alias';
-import { exclude } from '@bemedev/vitest-exclude';
+import { aliasTs } from '@bemedev/dev-utils/vitest-alias';
+import { exclude } from '@bemedev/dev-utils/vitest-exclude';
 import { defineConfig } from 'vitest/config';
 import tsconfig from './tsconfig.json';
 
@@ -16,11 +16,12 @@ export default defineConfig({
     fileParallelism: false,
     globals: true,
     logHeapUsage: true,
+    typecheck: {
+      enabled: true,
+    },
     coverage: {
       enabled: true,
-      extension: 'ts',
       reportsDirectory: '.coverage',
-      all: true,
       provider: 'v8',
     },
   },